+++ src/Core/Credentials/Scope.php
@@ -16,11 +16,12 @@ class Scope
     /**
      * @var string[]
      */
-    protected $availableScope = [
+    protected array $availableScope = [
         'app',
         'bizproc',
         'calendar',
         'call',
+        'catalog',
         'contact_center',
         'crm',
         'delivery',
@@ -60,7 +61,7 @@ class Scope
     /**
      * @var array
      */
-    protected $currentScope = [];
+    protected array $currentScope = [];
 
     /**
      * Scope constructor.
@@ -71,8 +72,7 @@ class Scope
      */
     public function __construct(array $scope = [])
     {
-        $scope = array_map('strtolower', $scope);
-        array_unique($scope);
+        $scope = array_unique(array_map('strtolower', $scope));
         foreach ($scope as $item) {
             if (!in_array($item, $this->availableScope, true)) {
                 throw new UnknownScopeCodeException(sprintf('unknown application scope code - %s', $item));