+++ src/Services/CRM/Product/Service/Batch.php
@@ -21,10 +21,11 @@ class Batch extends AbstractBatchService
      * batch list method
      *
      * @param array{
-     *                         ID?: int,
+     *                         ID?: string
+     *                         } $order
      *
      * @param array{
-     *                         ID?: int,
+     *                         ID?: int
      *                         } $filter
      * @param array    $select = ['ID','CATALOG_ID','PRICE','CURRENCY_ID','NAME','CODE','DESCRIPTION','DESCRIPTION_TYPE','ACTIVE','SECTION_ID','SORT','VAT_ID','VAT_INCLUDED','MEASURE','XML_ID','PREVIEW_PICTURE','DETAIL_PICTURE','DATE_CREATE','TIMESTAMP_X','MODIFIED_BY','CREATED_BY']
      * @param int|null $limit