+++ tests/Unit/Services/Main/MainServiceBuilderTest.php
@@ -7,6 +7,7 @@ namespace Bitrix24\SDK\Tests\Unit\Services\Main;
 use Bitrix24\SDK\Services\Main\MainServiceBuilder;
 use Bitrix24\SDK\Services\ServiceBuilder;
 use Bitrix24\SDK\Tests\Unit\Stubs\NullBatch;
+use Bitrix24\SDK\Tests\Unit\Stubs\NullBulkItemsReader;
 use Bitrix24\SDK\Tests\Unit\Stubs\NullCore;
 use PHPUnit\Framework\TestCase;
 use Psr\Log\NullLogger;
@@ -29,9 +30,13 @@ class MainServiceBuilderTest extends TestCase
         $this::assertTrue(true);
     }
 
-
     public function setUp(): void
     {
-        $this->serviceBuilder = (new ServiceBuilder(new NullCore(), new NullBatch(), new NullLogger()))->getMainScope();
+        $this->serviceBuilder = (new ServiceBuilder(
+            new NullCore(),
+            new NullBatch(),
+            new NullBulkItemsReader(),
+            new NullLogger()
+        ))->getMainScope();
     }
 }